+++ backend/attendance/src/main/java/org/egov/config/AttendanceServiceConfiguration.java
@@ -1,5 +1,6 @@
 package org.egov.config;
 
+import java.util.Map;
 import java.util.TimeZone;
 import jakarta.annotation.PostConstruct;
 
@@ -103,6 +104,9 @@ public class AttendanceServiceConfiguration {
     @Value("${attendance.register.open.search.enabled.roles}")
     private String registerOpenSearchEnabledRoles;
 
+    @Value("${attendance.log.open.search.enabled:false}")
+    private boolean logOpenSearchEnabled;
+
     //Individual servcie
     @Value("${works.individual.host}")
     private String individualHost;
@@ -111,6 +115,9 @@ public class AttendanceServiceConfiguration {
 
     @Value("${attendance.register.first.staff.insert.enabled:true}")
     private Boolean registerFirstStaffInsertEnabled;
+
+    @Value(("${attendance.register.first.owner.staff.enabled:false}"))
+    private boolean registerFirstOwnerStaffEnabled;
 
     //HRMS Service
     @Value("${egov.hrms.host}")
@@ -138,6 +145,26 @@ public class AttendanceServiceConfiguration {
     @Value("${project.staff.attendance.topic}")
     private String projectStaffAttendanceTopic;
 
+    @Value("${egov.boundary.host}")
+    private String boundaryServiceHost;
+
+    @Value("${egov.boundary.search.url}")
+    private String boundarySearchUrl;
+
+    @Value("${attendance.register.search.check.project.enabled:false}")
+    private Boolean attendanceRegisterProjectSearchEnabled;
+
+    @Value("${attendance.register.review.status.enabled:false}")
+    private Boolean attendanceRegisterReviewStatusEnabled;
+
+    @Value("${attendance.register.review.status.init.value}")
+    private String attendanceRegisterReviewStatusValue;
+
+    @Value("${attendance.register.boundary.search.enabled:false}")
+    private Boolean attendanceRegisterBoundarySearchEnabled;
+
+    @Value("#{${attendance.register.status.map}}")
+    private Map<String, String> attendanceRegisterStatusMap;
 }