+++ backend/attendance/src/main/java/org/egov/kafka/ProjectStaffConsumer.java
@@ -17,6 +17,7 @@ import org.egov.util.ProjectStaffUtil;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.messaging.handler.annotation.Header;
@@ -29,6 +30,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Component
+@ConditionalOnProperty(name = "project.staff.attendance.register.enabled", havingValue = "true", matchIfMissing = true)
 @Slf4j
 public class ProjectStaffConsumer {