+++ backend/attendance/src/main/java/org/egov/service/OrganisationContactDetailsStaffUpdateService.java
@@ -46,12 +46,12 @@ public class OrganisationContactDetailsStaffUpdateService {
             while (true){
                 AttendanceRegisterSearchCriteria attendanceRegisterSearchCriteria =
                     AttendanceRegisterSearchCriteria.builder().tenantId(tenantId).staffId(oldContact.getIndividualId()).limit(configuration.getAttendanceRegisterDefaultLimit()).offset(offSet).build();
-            List<AttendanceRegister> attendanceRegisterList = attendanceRegisterService.searchAttendanceRegister(requestInfoWrapper, attendanceRegisterSearchCriteria);
+            List<AttendanceRegister> attendanceRegisterList = attendanceRegisterService.searchAttendanceRegister(requestInfoWrapper, attendanceRegisterSearchCriteria).getAttendanceRegister();
             if(CollectionUtils.isEmpty(attendanceRegisterList)) {
                 try {
                     String userUuid = individualServiceUtil.getIndividualDetails(Collections.singletonList(oldContact.getIndividualId()), requestInfoWrapper.getRequestInfo(), tenantId).get(0).getUserUuid();
                     attendanceRegisterSearchCriteria = AttendanceRegisterSearchCriteria.builder().tenantId(tenantId).staffId(userUuid).limit(configuration.getAttendanceRegisterDefaultLimit()).offset(offSet).build();
-                    attendanceRegisterList = attendanceRegisterService.searchAttendanceRegister(requestInfoWrapper, attendanceRegisterSearchCriteria);
+                    attendanceRegisterList = attendanceRegisterService.searchAttendanceRegister(requestInfoWrapper, attendanceRegisterSearchCriteria).getAttendanceRegister();
                 }catch (Exception e){
                     log.error(e.toString());
                 }