+++ backend/attendance/src/main/resources/db/migration/main/V20250103104300__alter_attendance_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE eg_wms_attendance_register RENAME COLUMN paymentStatus TO reviewStatus;