+++ backend/expense/src/main/java/org/egov/digit/expense/repository/BillRepository.java
@@ -27,10 +27,14 @@ public class BillRepository {
 		this.searchBillRowMapper = searchBillRowMapper;
 	}
 
-	public List<Bill> search(BillSearchRequest billSearchRequest){
+	public List<Bill> search(BillSearchRequest billSearchRequest, boolean isValidationSearch){
 		
 		List<Object> preparedStatementValues = new ArrayList<>();
+<<<<<<< HEAD
 		String queryStr = queryBuilder.getBillQuery(billSearchRequest, preparedStatementValues, false);
+=======
+		String queryStr = queryBuilder.getBillQuery(billSearchRequest, preparedStatementValues, false, isValidationSearch);
+>>>>>>> 504a89d592593471db1fd567ee4faf870546941e
         return jdbcTemplate.query(queryStr, preparedStatementValues.toArray(), searchBillRowMapper);
 	}